+++ example/cli/src/app/app.module.ts
@@ -3,18 +3,35 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 import { TreeModule } from 'angular-tree-component';
+import { Router, Route, RouterModule } from '@angular/router';
 
 import { AppComponent } from './app.component';
+import { BasicTreeComponent } from './basictree/basictree.component';
+import { FullTreeComponent } from './fulltree/fulltree.component';
+
+const routes: Route[] = [
+  {
+    path: '',
+    component: FullTreeComponent
+  },
+  {
+    path: 'basic',
+    component: BasicTreeComponent
+  }
+];
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    BasicTreeComponent,
+    FullTreeComponent
   ],
   imports: [
     BrowserModule,
     FormsModule,
     HttpModule,
-    TreeModule
+    TreeModule,
+    RouterModule.forRoot(routes, { useHash: true })
   ],
   providers: [],
   bootstrap: [AppComponent]