+++ lib/components/tree-node-children.component.ts
@@ -14,12 +14,11 @@ import { deprecatedSelector } from '../deprecated-selector';
       <div [class.tree-children]="true"
           [class.tree-children-no-padding]="node.options.levelPadding"
           *ngIf="node.isExpanded">
-        <div *ngIf="node.children">
-          <tree-node-collection
-            [nodes]="node.children"
-            [templates]="templates">
-          </tree-node-collection>
-        </div>
+        <tree-node-collection
+          *ngIf="node.children"
+          [nodes]="node.children"
+          [templates]="templates">
+        </tree-node-collection>
         <tree-loading-component
           [style.padding-left]="node.getNodePadding()"
           class="tree-node-loading"